# api/swagger.yaml
swagger: "2.0"
info:
  description: >
    The River Endpoint Discovery Service (REDS) manages the endpoint discovery, configuration
    and geolocation of Redfish-enabled BMCs on River compute nodes. It does so through a
    combination of methods:
    watching switches for new MAC addresses via SNMP and booting a "discovery ramdisk".
   
    This API assumes that you have a general familiarity with SNMP, IPMI, and Redfish protocols and
    the operation of network switches.
    
    REDS involves the following tasks:
 
    * **Endpoint discovery**: This process determines that a new Redfish Endpoint (IPMI-capable BMC)
    is available on the network. 
    
    * **Configuration**: This process sets the minimum information for the node being discovered
    so that it can interact with the rest of the system. This typically consists of information
    such as credentials.
    
    * **Geolocation**: This process assigns a location-based identity to the node (typically
    based on the network wiring). It assigns an xname to the node based on its physical
    location and adds the node to the Hardware State Manager (HSM) service.
    
    
    The SNMP scanning performs the discovery and geolocation components. SNMP scanning
    allows the service to observe when a new River compute node is attached to the network.
    It also detects which port on which switch the new hardware is attached to. This is
    correlated with a pre-built map of the system to determine the logical name (xname) that
    should be associated with the newly detected BMC.
    
    
    When the new node first powers on, it is unknown to the rest of the system and boots to
    the discovery ramdisk. The discovery ramdisk handles bootstrapping the BMC with the correct
    username, password and other configuration. It does not assign an IP to the node, as the
    logical identity (xname) is not known. It does, however, collect the IP in use, meaning
    assignment could be carried out via DHCP or other similar mechanism.
    
    
    ## Resources
    
    ### /credentials
    
    Generate new credentials for a specific BMC on a River compute node. This is intended
    for interaction between the ramdisk (booted on new hardware) and daemon (on the SMS).
    It is not intended for human use.
    
    ### /discovery
    
    Accessed by nodes running the discovery and initialization initramfs
    to signal that the process is complete. This is intended for interaction between 
    the ramdisk (booted on new hardware) and daemon (on the non-compute node). It is not
    intended for human use.
    
    ## Workflows
    
    ### Add River Compute Nodes
    
    SLS is now used to add components. See "Expand System" in the SLS documentation.
    
    ### Remove River Compute Nodes
    
    SLS is now used to remove components. See "Remove Hardware" in the SLS documentation.
    
  version: "0.1"
  title: "River Endpoint Discovery Service"
  license:
    name: "Cray Proprietary"
host: "sms"
basePath: "/apis/reds/v1"
tags:
- name: "Credentials"
  description: "Get or change credentials for node BMCs"
- name: "Discovery"
  description: "Access to discovery functions"
- name: "Service Info"
  description: "Service information APIs such as readiness and liveness."
schemes:
- "https"
produces:
  - application/json
consumes:
  - application/json
paths:
  /readiness:
    get:
      tags:
        - Service Info
      summary: Kubernetes readiness endpoint to monitor service health
      x-private: true
      description: >-
        The `readiness` resource works in conjunction with the Kubernetes readiness
        probe to determine when the service is no longer healthy and able to respond
        correctly to requests.  Too many failures of the readiness probe will result
        in the traffic being routed away from this service and eventually the service
        will be shut down and restarted if in an unready state for too long.


        This is primarily an endpoint for the automated Kubernetes system.
      operationId: readiness_get
      responses:
        "204":
          description: >-
            [No Content](http://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html#sec10.2.5)
            Network API call success
        "503":
          description: "The service is unhealthy and not ready."
        default:
          description: "Unexpected error."

  /liveness:
    get:
      tags:
        - Service Info
      summary: Kubernetes liveness endpoint to monitor service health
      x-private: true
      description: >-
        The `liveness` resource works in conjunction with the Kubernetes liveness
        probe to determine when the service is no longer responding to
        requests.  Too many failures of the liveness probe will result in the
        service being shut down and restarted.


        This is primarily an endpoint for the automated Kubernetes system.
      operationId: liveness_get
      responses:
        "204":
          description: >-
            [No Content](http://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html#sec10.2.5)
            Network API call success
        default:
          description: "Unexpected error."

  /credentials:
    post:
      tags:
      - "Credentials"
      - "cli_ignore"
      operationId: "credsByMAC_post"
      summary: "Generate new credentials for the BMC"
      description: "Generate new credentials for the BMC as identified by the request payload.
      This is a POST request since credentials are generated."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: body
          name: addesses
          schema:
            $ref: "#/definitions/SystemAddresses"
      responses:
        200:
          description: "New credentials generated and included in the response"
          schema:
            $ref: "#/definitions/BMCCredentials"
        405:
          description: "Invalid input"
    
  /discovery:
    put:
      tags:
      - "Discovery"
      - "cli_ignore"
      operationId: "discovery_put"
      summary: "Signal completion of initialization and discovery"
      description: "This resource is accessed by nodes running the discovery and initialization
      initramfs to signal that the process is complete."
      parameters:
        - in: body
          name: macAddesses
          schema:
            $ref: "#/definitions/SystemAddresses"
      responses:
        200:
          description: OK
        405:
          description: "Invalid input"

definitions:
  BMCCredentials:
    description: "An object which contains the credentials for a BMC."
    type: "object"
    properties:
      username:
        description: "The username portion of the BMC credentials."
        type: "string"
      password:
        description: "The password portion of the BMC credentials."
        type: "string"
  SystemAddresses:
    description: "An object to convey all the addresses assigned to a BMC"
    type: "object"
    properties:
      addresses:
        description: |
          An array of addresses on a per-interface basis.  This includes MAC addresses and
          IP addresses
          
          
          No correlation must be maintained between the various items in the addresses array
          after they arrive at the server. While they are all on the same physical machine
          and should be set to identical credentials, only one BMC (and therefore only one
          MAC address) should be configured as the BMC. The rest will be ignored. Additionally,
          it is recommended that the MAC address be used as a primary key in any storage
          configuration, as it is a unique, non-volatile identifier for the device.
        type: "array"
        items:
          type: "object"
          properties:
            macAddress:
              type: "string"
              description: "A MAC address, encoded in lowercase hexadecimal, without separators."
            IPAddresses:
              type: "array"
              items:
                type: "object"
                properties:
                  addressType:
                    type: "string"
                    enum: ["IPv4", "IPv6"]
                    description: "The type of the IP address conveyed.  Either IPv4 or IPv6."
                  address:
                    type: "string"
                    description: "The address itself."